fix: updating board card html after position changes
Before this change, the card HTML wasn't being updated
so the card link that points to it's detail form was still using
the old group_id.

The test was modified to fetch the new HTML element returned
by the turbo_stream response. Otherwise capybara would
show an error: stale element reference: stale element not found in the current frame

This caused a very nasty bug that you could reproduce by:

- Entering a board with multiple columns
- Move a card to another column
- Remove the older column
- Click in a card
- You will see a "Content missing" error in the screen because Rails
is responding with an error

The route doesn't need to contain the grouping id so this will be refactored later