Updates profile specs to use a city that serves sushi and doesn't use summer time
The profile spec started failing because it was asking capybara
to select an option with "Rome (GMT+01:00)" but the summer timer just
started today and the correct option now is "Rome (GMT+02:00)".

The test was failing because it couldn't find the "Rome (GMT+01:00)"
option as it's automatically updated.

So I'm changing the test to select an option with a place that doesn't adopt
summer time and have amazing food: Tokyo